refactor: made a lot of changes
