+++ routes/index.js
@@ -1,25 +1,28 @@
 const express = require('express')
 const router = express.Router()
-const userController = require('../controllers/user-controller')
-const admin = require('./modules/admin')
-const tweets = require('./modules/tweets')
-const { apiErrorHandler } = require('../middleware/error-handler')
 
-router.use('/api/tweets', tweets)
+const admin = require('./modules/admin')
+
+const userController = require('../controllers/apis/user-controller')
+const tweetContorller = require('../controllers/apis/tweet-controller')
+
+const { apiErrorHandler } = require('../middleware/error-handler')
 
 // api/admin
 router.use('/api/admin', admin)
 
 // api/users
-router.get('/api/users', userController.getUsers)
-router.get('/api/signup', userController.signUpPage)
+// router.get('/api/users', userController.getUsers)
+// router.get('/api/signup', userController.signUpPage)
 router.post('/api/signup', userController.signUp)
 
 // api/tweets
+router.get('/api/tweets/:tweet_id', tweetContorller.getTweet)
+router.get('/api/tweets', tweetContorller.getTweets)
 
 // router.get("/restaurants", restController.getRestaurants);
 // router.use("/", (req, res) => res.redirect("/restaurants"));
 
-router.use('/api/tweets', tweets)
 router.use('/', apiErrorHandler)
+
 module.exports = router
