+++ routes/modules/admin.js
@@ -1,8 +1,8 @@
 const express = require('express')
 const router = express.Router()
-const adminController = require('../../controllers/admin-controller')
+const adminController = require('../../controllers/apis/admin-controller')
 
-router.post('/signIn', adminController.signIn)
+// router.post('/signIn', adminController.signIn)
 router.get('/users', adminController.getUsers)
 router.delete('/tweets/:id', adminController.deleteTweet)
 router.get('/tweets', adminController.getTweets)
